cleanup progressdots component
