+++ components/carousel/project-data.tsx
